refactor role service from auth service
