+++ medical-portal/src/UI/src/app/features/auth/enums/role.enum.ts
@@ -1,9 +1,3 @@
-export enum IdentityProvider {
-  BCSC = 'bcsc',
-  IDIR = 'idir',
-  PHSA = 'phsa',
-}
-
 export enum Role {
   Practitioner = 'PRACTITIONER',
   Enrolled = 'DMFT_ENROLLED',
