--- medical-portal/src/UI/src/app/features/auth/services/access-token.service.spec.ts
@@ -1,21 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { provideAutoSpy } from 'jest-auto-spies';
-import { KeycloakService } from 'keycloak-angular';
-
-import { AccessTokenService } from './access-token.service';
-
-describe('AccessTokenService', () => {
-  let service: AccessTokenService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AccessTokenService, provideAutoSpy(KeycloakService)],
-    });
-    service = TestBed.inject(AccessTokenService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
