--- medical-portal/src/UI/src/app/modules/permissions/permissions.service.spec.ts
@@ -1,29 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-
-import { provideAutoSpy } from 'jest-auto-spies';
-
-import { AccessTokenService } from '@app/features/auth/services/access-token.service';
-
-import { PermissionsService } from './permissions.service';
-
-describe('PermissionsService', () => {
-  let service: PermissionsService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [RouterTestingModule],
-      providers: [
-        PermissionsService,
-        provideAutoSpy(AccessTokenService),
-        provideAutoSpy(PermissionsService),
-      ],
-    });
-
-    service = TestBed.inject(PermissionsService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
